+++ src/lib/components/page/Print.svelte
@@ -1,9 +1,25 @@
 <script lang="ts">
     import Chapter from './Chapter.svelte';
     import { getEdition } from './Contexts';
+    import Parser from '$lib/models/chapter/Parser';
 
     let edition = getEdition();
+
+    $: title = $edition
+        ? [
+              $edition.getTitle(),
+              `${$edition.getEditionLabel()} Edition`,
+              Parser.parseFormat($edition, $edition.getAuthors().at(0) ?? '').toText() +
+                  ($edition.getAuthors().length > 1 ? ' et al.' : ''),
+          ]
+              .filter(Boolean)
+              .join(' - ')
+        : undefined;
 </script>
+
+<svelte:head>
+    {#if title}<title>{title}</title>{/if}
+</svelte:head>
 
 {#if $edition}
     {#each $edition